\name{tools4uplift-package}
\alias{tools4uplift-package}
\alias{tools4uplift}
\docType{package}
\title{
\packageTitle{tools4uplift}
}
\description{
\packageDescription{tools4uplift}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{tools4uplift}
\packageIndices{tools4uplift}

}
\author{
Mouloud Belbahri <mouloud.belbahri@gmail.com>,  Olivier Gandouet, Alejandro Murua, Vahid Partovi Nia
}
\keyword{uplift}
